Safari-611.1.10/LayoutTests/accessibility/html5-required-attribute-expected.txt


This tests that the required attribute is returned correctly through accessibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.isRequired is true
PASS accessibilityController.focusedElement.isRequired is false
PASS fileUpload.isRequired is true
PASS successfullyParsed is true

TEST COMPLETE
